fix: [CO-1348] prevent resource leaks in preview servlet
